Azure/vantage-cost-ingestion/requirements.txt
azure-functions
azure-identity
azure-monitor-ingestion
azure-storage-blob